Widen Register_Memory addr parameter and reorder MemoryWrite stores
Pi1MHz_Register_Memory took uint8_t addr, which made every caller
narrow its address expression at the call boundary. The (uint8_t)
casts at 31 call sites forced the compiler to compute each wrapped
index separately (uxtb + per-index add) instead of sharing one scaled
base. Widening the parameter to unsigned int and dropping the casts
lets the compiler compute the callback_table address once per run of
registrations: rambyte_emulator_init shrinks from 21 instructions to
11. No caller registers a run that wraps the page, so behaviour is
unchanged (the closest is rampage at 0xFD..0xFF).

In Pi1MHz_MemoryWrite, store the plain read-back byte before the
volatile VPU read-modify-write instead of after it. The store is
independent of the VPU access, so the CPU issues it during the VPU
load's result latency instead of stalling behind it; this also
matches the ordering MemoryWrite16/32 already use. The BBC now sees
the read-back byte a couple of CPU cycles before the VPU word update
rather than after it, which is invisible at 1MHz bus granularity.

Net: kernel.img -256 bytes, 2 cycles saved per FRED write callback.

+++ src/BeebSID/BeebSid.c
@@ -65,7 +65,7 @@ void BeebSID_emulator_init(uint8_t instance, uint8_t address)
     rpi_audio_init(beebsid_sample_rate);
 
     for (i = 0; i < 32u; i++) {
-        Pi1MHz_Register_Memory(WRITE_FRED, (uint8_t)(address + i), beebsid_write);
+        Pi1MHz_Register_Memory(WRITE_FRED, (address + i), beebsid_write);
     }
     Pi1MHz_Register_Poll(beebsid_poll);
 }
