+++ src/framebuffer/framebuffer.c
@@ -2471,8 +2471,8 @@ void fb_emulator_init(uint8_t instance, uint8_t address)
   Pi1MHz_MemoryWrite((uint32_t)(address+2), 0XFC);
   Pi1MHz_MemoryWrite((uint32_t)(address+3), 0x4c); // JMP
 
-  Pi1MHz_Register_Memory(WRITE_FRED, (uint8_t)(address + 4), Pi1MHz_EmulatedMemoryByte);
-  Pi1MHz_Register_Memory(WRITE_FRED, (uint8_t)(address + 5), Pi1MHz_EmulatedMemoryByte);
+  Pi1MHz_Register_Memory(WRITE_FRED, (address + 4), Pi1MHz_EmulatedMemoryByte);
+  Pi1MHz_Register_Memory(WRITE_FRED, (address + 5), Pi1MHz_EmulatedMemoryByte);
 
  // Pi1MHz_Register_Poll(fb_emulator_poll);
 }
